Tidied up and fixed unit tests
