+++ test/utils/url.unit.ts
@@ -1,5 +1,7 @@
 import { STOP_TYPE } from "../../src/utils/constants";
-import { urlUtils, sanitizeStopType } from "../../src/utils/url";
+import { urlUtils, sanitizeStopType, getPreviousPageQueryParamUrl } from "../../src/utils/url";
+import { OFFICER_FILING } from "../../src/types/page.urls";
+
 import { Request } from "express";
 
 describe("getCompanyNumberFromRequestParams", () => { 
@@ -164,5 +166,57 @@ describe("getCompanyNumberFromRequestParams", () => {
   it("should sanitize multiple stop types to select none", () => {
     const result = sanitizeStopType(["!pre-october-2009ABC", "pre-october-2009"]);
     expect(result).toBe("");
+  });
+});
+
+describe('getPreviousPageQueryParamUrl tests', () => {
+  it('should return the previous page url from headers', () => {
+    const previousPage = OFFICER_FILING + '/test'
+    const req = {
+      rawHeaders: [
+        'Host', 'localhost:3000',
+        'User-Agent', 'curl/7.64.1',
+        'Accept', '*/*',
+        'Referer', previousPage
+      ],
+      query: {
+      }
+    }
+    
+    expect(getPreviousPageQueryParamUrl(req as any)).toBe(previousPage)
+  });
+
+  it('should return the previous page url from headers if previousPage query param has multiple values', () => {
+    const previousPage = OFFICER_FILING + '/test'
+    const req = {
+      rawHeaders: [
+        'Host', 'localhost:3000',
+        'User-Agent', 'curl/7.64.1',
+        'Accept', '*/*',
+        'Referer', previousPage
+      ],
+      query: {
+        previousPage: [OFFICER_FILING, "http://example.com"]
+      }
+    }
+    
+    expect(getPreviousPageQueryParamUrl(req as any)).toBe(previousPage)
+  });
+
+  it('should return the previous page url from single previousPage query param', () => {
+    const previousPage = OFFICER_FILING + '/test'
+    const req = {
+      rawHeaders: [
+        'Host', 'localhost:3000',
+        'User-Agent', 'curl/7.64.1',
+        'Accept', '*/*',
+        'Referer', OFFICER_FILING
+      ],
+      query: {
+        previousPage: previousPage
+      }
+    }
+    
+    expect(getPreviousPageQueryParamUrl(req as any)).toBe(previousPage)
   });
 });
